fix: ID type fix in JpaRepositories

+++ harmoni-data/src/main/java/org/harmoniapp/harmonidata/repositories/AbsenceRepository.java
@@ -3,5 +3,5 @@ package org.harmoniapp.harmonidata.repositories;
 import org.harmoniapp.harmonidata.entities.Absence;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AbsenceRepository extends JpaRepository<Absence, Integer> {
+public interface AbsenceRepository extends JpaRepository<Absence, Long> {
 }
