+++ harmoni-data/src/main/java/org/harmoniapp/harmonidata/repositories/AvailabilityRepository.java
@@ -3,5 +3,5 @@ package org.harmoniapp.harmonidata.repositories;
 import org.harmoniapp.harmonidata.entities.Availability;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AvailabilityRepository extends JpaRepository<Availability, Integer> {
+public interface AvailabilityRepository extends JpaRepository<Availability, Long> {
 }
