+++ harmoni-data/src/main/java/org/harmoniapp/harmonidata/repositories/PredefineShiftRepository.java
@@ -3,5 +3,5 @@ package org.harmoniapp.harmonidata.repositories;
 import org.harmoniapp.harmonidata.entities.PredefineShift;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PredefineShiftRepository extends JpaRepository<PredefineShift, Integer> {
+public interface PredefineShiftRepository extends JpaRepository<PredefineShift, Long> {
 }
